Update and rename Department of Health (DoH) – Pass 1.csv to National Departments/Department of Health/Department of Health (DoH) – Pass 1.csv
#104  
Removed R's, "," and spaces and fixing dates. 
I renamed the file and put it in the correct location

I also split some of the columns, could you please check if the way in which I did it is alright.
There are two rows in the "Unit price" column that have descriptions, I am not sure wether or not that should be removed.

Thanks!